django/doccheck_service/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
python-dotenv==1.0.1
requests==2.31.0
drf-spectacular==0.27.2
django-cors-headers==4.3.0
